+++ stringmatch/__init__.py
@@ -1,7 +1,8 @@
 # flake8: noqa
+from .distance import *
 from .match import *
 from .ratio import *
 from .strings import *
 
 __title__ = "stringmatch"
-__version__ = "0.5.0"
+__version__ = "0.6.0"