+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This is a broad overview of the changes that have been made over the lifespan of this library.
 
+## v0.6.0 - ????-??-??
+
+- Added distance functions
+    - These are: distance and distance_list
+
 ## v0.5.0 - 2022-04-27
 
 - Removed scorer argument from functions, added it into `__init__` in both Match() and Ratio()